Port feedback API test to jest; remove old test infrastructure and modules
