+++ app/resources/v1/__tests__/translate.test.js
@@ -21,11 +21,10 @@ describe('get api/v1/translate', function () {
   it('makes a request for a translation file', () => {
     return request(app)
       .get('/api/v1/translate/en/main')
-      .expect('Content-Type', 'application/json; charset=utf-8')
-      .expect(200)
       .then((response) => {
-        const translation = response.body
-        expect(translation.dialogs.welcome.heading).toEqual('Welcome to Streetmix.')
+        expect(response.statusCode).toEqual(200)
+        expect(response.get('Content-Type').toLowerCase(), 'application/json; charset=utf-8')
+        expect(response.body.dialogs.welcome.heading).toEqual('Welcome to Streetmix.')
       })
   })
 })
