--- test/index.js
@@ -1,4 +1,0 @@
-process.env.NODE_ENV = 'test'
-
-// run tests
-require('./api.v1.feedback.js')
